Launches files update
- Rename to suggested names
- Added/Declared launch arguments for every launch files
