+++ racecar_bringup/launch/GRO525.launch.py
@@ -1,0 +1,59 @@
+import os
+
+from ament_index_python.packages import get_package_share_directory
+from launch import LaunchDescription
+from launch_ros.actions import Node
+from launch.actions import DeclareLaunchArgument, IncludeLaunchDescription
+from launch_xml.launch_description_sources import XMLLaunchDescriptionSource
+from launch.launch_description_sources import PythonLaunchDescriptionSource
+from launch.substitutions import LaunchConfiguration
+
+racecar_bringup_dir = get_package_share_directory("racecar_bringup")
+rosbridge_server_dir = get_package_share_directory("rosbridge_server")
+
+rosbridge_server_ld = IncludeLaunchDescription(
+    XMLLaunchDescriptionSource(
+        f"{rosbridge_server_dir}/launch/rosbridge_websocket_launch.xml"
+    ),
+    launch_arguments={"port": LaunchConfiguration("port")}.items(),
+)
+
+web_video_server_node = Node(
+    package="web_video_server",
+    executable="web_video_server",
+    name="web_video_server",
+    output="screen",
+    parameters=[{"address": LaunchConfiguration("host_address")}],
+)
+
+def generate_launch_description():
+    return LaunchDescription([
+        # Declare launch arguments
+        DeclareLaunchArgument('start_robotStatePublisher', default_value='False'),
+        DeclareLaunchArgument('start_arduinoBridge', default_value='True'),
+        DeclareLaunchArgument('start_arduinoSensor', default_value='False'),
+        DeclareLaunchArgument('start_lidar', default_value='False'),
+        DeclareLaunchArgument('start_camera', default_value='True'),
+        DeclareLaunchArgument('start_magwick', default_value='False'),
+        DeclareLaunchArgument('start_kalma', default_value='False'),
+        DeclareLaunchArgument("host_address", 
+                              default_value="10.42.0.1", 
+                              description="Address of the Raspberry Pi"),
+        DeclareLaunchArgument("port", 
+                              default_value="9090", 
+                              description="Port for rosbridge websocket"),
+        IncludeLaunchDescription(
+            PythonLaunchDescriptionSource([os.path.join(racecar_bringup_dir, 'launch', 'bringup.launch.py')]),
+            launch_arguments={'start_robotStatePublisher': LaunchConfiguration('start_robotStatePublisher'),
+                              'start_arduinoBridge': LaunchConfiguration('start_arduinoBridge'),
+                              'start_arduinoSensor': LaunchConfiguration('start_arduinoSensor'),
+                              'start_lidar': LaunchConfiguration('start_lidar'),
+                              'start_camera': LaunchConfiguration('start_camera'),
+                              'start_magwick': LaunchConfiguration('start_magwick'),
+                              'start_kalma': LaunchConfiguration('start_kalma')}.items()
+        ),
+
+            rosbridge_server_ld,
+            web_video_server_node,
+        ]
+    )
